Updated to a working version of BEER_curve
There is one small inconsistency in the code that I have opted to ignore: the planetary phase curve continues to evolve and contribute during the bottommost part of the eclipse, even when it should actually be zeroed out.

As far as I can tell, there is no self-consistent way to have the phase curve zero out and to ALSO have a variable eclipse depth parameter. However, since the phase curve varies by only about 1 ppm for typical parameters during the eclipse (much below the uncertainties), I have opted to leave this small inconsistency in.

A better solution would be to use a full planetary phase curve calculator, such as STARRY. Maybe in the future.